refactor: split & apply getting api url & api calls
